fix: dto와 query의 반환값이 타입이 다른 문제 해결

+++ project/back/health/src/main/java/com/ssafy/health/domain/crew/dto/analysis/MaxScoresDto.java
@@ -7,6 +7,6 @@ import lombok.Data;
 @AllArgsConstructor
 public class MaxScoresDto {
 
-    private float basicScore;
-    private float activityScore;
+    private Double basicScore;
+    private Double activityScore;
 }
